scripts/requirements.txt
python-amazon-paapi>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
